+++ src/ui/Input.js
@@ -6,6 +6,7 @@ require('jquery-placeholder');
  * @param {HTMLElement} input - 目标元素
  * @param {Object} options - 参数
  * @param {String} options.theme - 主题
+ * @param {String} options.customClass - 自定义 CSS class
  * @param {Boolean} options.disabled - 禁用，默认 false
  * @param {Function} options.onEnter - 按回车回调
  */
@@ -14,7 +15,8 @@ function Input(input, options) {
     this.$main = $(this.main);
 
     var defaultOption = {
-        theme: bizui.theme
+        theme: bizui.theme,
+        customClass: ''
     };
     this.options = $.extend(defaultOption, options || {});
     this.init(this.options);
@@ -33,7 +35,7 @@ Input.prototype = {
      * @private
      */
     init: function(options) {
-        this.$main.addClass(defaultClass);
+        this.$main.addClass(defaultClass + ' ' + options.customClass);
 
         this.$main.placeholder();
 
@@ -82,7 +84,7 @@ Input.prototype = {
      * 销毁
      */
     destroy: function() {
-        this.$main.removeClass(defaultClass + ' ' + disableClass);
+        this.$main.removeClass([defaultClass, this.options.customClass, disableClass].join(' '));
         this.$main.off('keydown.bizInput')
             .off('mouseover.bizInput')
             .off('mouseout.bizInput')